ocr/requirements.txt
fastapi
uvicorn
Pillow
pdf2image
onnxruntime
rapidocr_onnxruntime
